requirements.txt
tensorflow>=2.12.0
opencv-python
numpy
matplotlib
lxml
Pillow